requirements.txt
Flask==2.3.3
markdown==3.5.1
PyYAML==6.0.1
python-dotenv==1.0.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
duckduckgo-search==3.9.6
google-generativeai>=0.7.0
graphviz==0.20.1
